requirements.txt
aiodns==3.2.0
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
babel==2.16.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
colorama==0.4.6
fluent.runtime==0.4.0
fluent.syntax==0.19.0
frozenlist==1.5.0
i18n==0.2
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
pillow==11.0.0
propcache==0.2.0
py==1.11.0
pycares==4.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.0
pytz==2024.2
structlog==24.4.0
typing_extensions==4.12.2
ujson==5.10.0
yarl==1.17.0